% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatedbars.R
\name{animatedbars}
\alias{animatedbars}
\title{Generate animated bar chart for counts by subject}
\usage{
animatedbars(data, subjectcol, yearcol, countcol, rankcut, gifname,
  gifwidth, gifheight, gifspeed, framespersecond)
}
\arguments{
\item{data}{A data frame containing the variables of interest.}

\item{subjectcol}{is a quoted reference to the column in your \code{data} that contains your subject areas/codes}

\item{yearcol}{is a quoted reference to the column with your time parameter (semester, year, etc.)}

\item{countcol}{is a quoted reference to the column in your \code{data} that contains the enrollment count per \code{subjectcol} and \code{yearcol} - this means you will need to do your counting and filtering beforehand}

\item{rankcut}{is an unquoted number that determines how many bars you would like (top 10, top 20, etc)}

\item{gifname}{is a quoted name for your gif, if you chose to write the animation to a .gif file}

\item{gifwidth}{is an unquoted number of pixels used to determine the width of the animation (if left blank it will default to 800 pixels)}

\item{gifheight}{is an unquoted number of pixels used to determine the height of the animation (if left blank it will default to 800 pixels)}

\item{gifspeed}{is an unquoted number of seconds used to determine the length of the total animation (default is 15)}

\item{framespersecond}{determines how many frames per second (if left blank it will default to 15 fps)}
}
\description{
Generate animated bar chart for counts by subject
}
