% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diexcel.R
\name{direport}
\alias{direport}
\title{Generate Grading Equity Report for Instructor and All Faculty by course taught}
\usage{
direport(data, subjectcol, subject, coursecol, racecol, gendercol,
  gradecol, limit)
}
\arguments{
\item{data}{A data frame containing the variables of interest. Must include a column for \code{course}, \code{coursename},\code{characteristic},\code{grades}.}

\item{subjectcol}{is a quoted reference to the column in your \code{data} that contains your subject codes/course designators (ENG, MATH, PHYS, etc.)}

\item{subject}{is a quoted reference to the subject code/course designator you would like to generate the DI report for.}

\item{coursecol}{is a quoted reference to the column containing the identifiers for your courses (e.g., ENGWR 300, MATH 100, etc.)}

\item{racecol}{is a quoted reference to the column in your \code{data} that contains the race/ethnicity indicator for each student.}

\item{gendercol}{is a quoted reference to the column in your \code{data} that contains the gender indicators.}

\item{gradecol}{is a quoted reference to the column in your \code{data} that contains the letter grades for each student has earned. IMPORTANT!! this must be formatted A,B,C,D,F,W (where W is withdrawl and similar) and other letter grades do not contain +/- grades. Grades must be capitalized.}

\item{limit}{is an unquoted integer specifying the lower limit of the count required to be included (e.g., limit=10 will replace data for any group with less than 10 students with "---". This is optional and will default to include all groups if limit is not specified. Use this for masking groups with small N.}
}
\description{
Generate Grading Equity Report for Success (A,B,C grades), Mastery (A,B grades), and Drops (W grades) for identified instructor and for all instructors who have taught each course that the identified instructor has taught.
}
\examples{
direport(student_data,"Subject","English","Course","Race","Gender","Official_Grade",10)
}
