% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equityrates.R
\name{equitysheet}
\alias{equitysheet}
\title{Generate Grading Equity Report for Instructor and All Faculty by course taught}
\usage{
equitysheet(data, coursecol, instructorcol, instructorid, gradecol,
  racecol, gendercol, limit, footer, savesheet)
}
\arguments{
\item{data}{A data frame containing the variables of interest. Must include a column for \code{course}, \code{coursename},\code{characteristic},\code{grades}.}

\item{coursecol}{is a quoted reference to the column in your \code{data} that contains your course titles}

\item{instructorcol}{is a quoted reference to the column containing the identifiers for your instructors (e.g., instructor name, employee id, etc.)}

\item{instructorid}{is a quoted reference to the identifier of the instructor you wish to examine (e.g., an instructor name or ID number depending on what you selected as your \code{instructorcol} )}

\item{gradecol}{is a quoted reference to the column in your \code{data} that contains the letter grades for each student has earned. IMPORTANT!! this must be formatted A,B,C,D,F,W (where W is withdrawl and similar) and other letter grades do not contain +/- grades. Grades must be capitalized.}

\item{racecol}{is a quoted reference to the column in your \code{data} that contains the race/ethnicity indicator for each student.}

\item{gendercol}{is a quoted reference to the column in your \code{data} that contains the gender indicators.}

\item{limit}{is an unquoted integer specifying the lower limit of the count required to be included (e.g., limit=10 will replace data for any group with less than 10 students with "---". This is optional and will default to include all groups if limit is not specified. Use this for masking groups with small N.}

\item{footer}{is an unquoted integer of 1 if you would like there to be a footer that shows "Equity Report for: 'PROFESSOR NAME' - Generated: 'DATE'. If you don't want the footer, leave this blank/don't specify.}

\item{savesheet}{is a quoted "Y" for yes or "N" for no, defaulting to "N" if left blank. If "Y" then it will save the graph as a .png in your working directory with a file name leading with the \code{instructorname}. This allows for looping and generating graphs for an entire dept or all courses an instructor teaches.}
}
\description{
Generate Grading Equity Report for Success (A,B,C grades), Mastery (A,B grades), and Drops (W grades) for identified instructor and for all instructors who have taught each course that the identified instructor has taught.
}
\examples{
equitysheet(student_data,"Course","Instructor",
"Professor 1","Official_Grade","Race","Gender",10,1,"N")
}
