% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradedistrofaculty.R
\name{gradedistrofaculty}
\alias{gradedistrofaculty}
\title{Generate graph for course grade distribution by filter characteristic (e.g., grades by race)}
\usage{
gradedistrofaculty(data, coursecol, coursename, instructorcol,
  instructorname, filtercol, gradecol, graphsubtitle, limit, legend,
  leadingzero, savesheet)
}
\arguments{
\item{data}{A data frame containing the variables of interest.}

\item{coursecol}{is a quoted reference to the column in your \code{data} that contains your course titles (alernatively you could also point to a column such as SUBJECT CODE and see distributions by subject rather than course).}

\item{coursename}{is a quoted name of a course that you would like to see the grade distribution for (e.g. ENG 101).}

\item{instructorcol}{is a quoted reference to the column that contains your instructor ID data (e.g., instructor name or instructor employee ID number).}

\item{instructorname}{is the quoted name of the instructor you want to generate distribution for.}

\item{filtercol}{is a quoted reference to the column in which your filtering characteristic for the students is found (e.g., Race, Instruction.Mode, Gender, CourseTime, etc.)}

\item{gradecol}{is a quoted reference to the column in your \code{data} that contains the letter grades}

\item{graphsubtitle}{is a quoted subtitle that will appear in smaller font just beneath title of the graph (something such as the date range for the data (Fall 16 to Spring 19) might be appropriate) - this is option and will auto fill in the graph if not specified}

\item{limit}{is an unquoted integer specifying the lower limit of the count required to be included (e.g., limit=10 will remove all groups that do not have a total of at least 10 entries...so if filtering characteristic is Race, it would remove all racial groups with less than 10 students in the course) - this is optional and will default to including all groups if limit is not specified}

\item{legend}{is a quoted title for your graph legend. If you don't include anything it will remain blank. Example might be "Race/Ethnicity" or "Gender"}

\item{leadingzero}{is an unquoted integer that allows you to set the number of leading zeros for ID numbers...R will read without zeros so in your .csv an ID might be 0000123 but R may just show 123 so setting \code{leadingzero} to 7 will ensure that the number is padded with zeros until it is 7 digits long to reflect 0000123.}

\item{savesheet}{is a quoted "Y" for yes or "N" for no, defaulting to "N" if left blank. If "Y" then it will save the graph as a .png in your working directory with a file name leading with the \code{instructorname}. This allows for looping and generating graphs for an entire dept or all courses an instructor teaches.}
}
\description{
Generate graph for grade distribution in course by filter characteristic
}
\examples{
gradedistrofaculty(student_data,"Course","English 101","Instructor","Professor 1",
"Race","Official_Grade","Grading Distribution for English 101",
10,"Race (# students)",0,"N")
}
