% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradingvar.R
\name{gradevar}
\alias{gradevar}
\title{Generate grading variability for success, mastery, drop/dfw by course and instructor}
\usage{
gradevar(data, coursecol, coursename, instructorcol, instructorname,
  successcol, masterycol, dropcol, indicator, limit, graphtitle,
  leadingzero, savesheet)
}
\arguments{
\item{data}{A data frame containing the variables of interest.}

\item{coursecol}{is a quoted reference to the column in your \code{data} that contains your course titles (alernatively you could also point to a column such as SUBJECT CODE and see distributions by subject rather than course).}

\item{coursename}{is a quoted name of a course that you would like to see the grade distribution for (e.g. ENG 101).}

\item{instructorcol}{is a quoted reference to the column that contains your instructor ID data (e.g., instructor name or instructor employee ID number).}

\item{instructorname}{is the quoted name of the instructor you want to generate distribution for.}

\item{successcol}{is a quoted reference to the column in which course success is indicated at 1 or 0}

\item{masterycol}{is a quoted reference to the column in which course mastery (usually grade A or B) is indicated at 1 or 0}

\item{dropcol}{is a quoted reference to the column in which course drops (either W, of D,F,W, whatever makes sense for your institution) is indicated at 1 or 0}

\item{indicator}{is an unquoted integer that indicates your positive indicator (1 = yes, usually)}

\item{limit}{is an unquoted integer specifying the lower limit of the count required to be included (e.g., limit=10 will remove all groups that do not have a total of at least 10 entries...so if filtering characteristic is Race, it would remove all racial groups with less than 10 students in the course) - this is optional and will default to including all groups if limit is not specified}

\item{graphtitle}{is a quoted title that will appear on the graph - this is optional and will auto fill in the graph title if not specified}

\item{leadingzero}{is an unquoted integer that allows you to set the number of leading zeros for ID numbers...R will read without zeros so in your .csv an ID might be 0000123 but R may just show 123 so setting \code{leadingzero} to 7 will ensure that the number is padded with zeros until it is 7 digits long to reflect 0000123.}

\item{savesheet}{is a quoted "Y" for yes or "N" for no, defaulting to "N" if left blank. If "Y" then it will save the graph as a .png in your working directory with a file name leading with the \code{instructorname}. This allows for looping and generating graphs for an entire dept or all courses an instructor teaches.}
}
\description{
Generate grading variability for success, mastery, drop/dfw by course and instructor
}
\examples{
gradevar(student_data,"Course","English 101","Instructor","Professor 3",
"success","mastery","drop",1,10,"Grading variability",0,"N")
}
