% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_data.R
\docType{data}
\name{student_data}
\alias{student_data}
\title{Sample Student Data for Analysis}
\format{An object of class \code{data.frame} with 9000 rows and 11 columns.}
\usage{
student_data
}
\description{
A \code{data.frame} with 11 variables and 9000 observations
}
\details{
Generate student_data sample data
}
\keyword{datasets}
